Change levels type to double; missing grid adjustement in squared curvature term

+++ lab/example/input/include/InputData.h
@@ -36,7 +36,7 @@ namespace SCaBOliC
                 iterations = 10;
             }
 
-            int levels;
+            double levels;
             ODRModel::LevelDefinition ld;
             ODRModel::NeighborhoodType nt;
